refactor: move availability into its own file, prep for time window
