+++ packages/skylarktv/src/mocks/airtable/utils.ts
@@ -1,14 +1,6 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { Record as AirtableRecord, Attachment, FieldSet } from "airtable";
-import { Airtables } from "../../types/airtable";
 import { Maybe } from "../../types";
-
-// Extract availability dimensions from request headers (case insensitive)
-export interface AvailabilityDimensions {
-  customerTypes: string[];
-  deviceTypes: string[];
-  regions: string[];
-}
 
 export type AirtableField = FieldSet[string];
 
@@ -228,151 +220,6 @@ export const mergeTranslatedContent = (
   });
 
   return mergedFields;
-};
-
-// Availability support functions
-export const checkAvailabilityMatch = (
-  availabilityRecord: AirtableRecord<FieldSet>,
-  requestedDimensions: AvailabilityDimensions,
-  airtableData: Airtables,
-): boolean => {
-  // Get the availability record's dimensions
-  let availabilityCustomers =
-    assertStringArray(availabilityRecord.fields.customers) || [];
-  let availabilityDevices =
-    assertStringArray(availabilityRecord.fields.devices) || [];
-  let availabilityRegions =
-    assertStringArray(availabilityRecord.fields.regions) || [];
-
-  // If no direct dimensions, check if this availability uses segments
-  const availabilitySegments =
-    assertStringArray(availabilityRecord.fields.segments) || [];
-
-  // If segments are used, resolve them to all dimension types
-  if (availabilitySegments.length > 0) {
-    availabilitySegments.forEach((segmentId) => {
-      const segment = airtableData.audienceSegments?.find(
-        (s) => s.id === segmentId,
-      );
-      if (segment) {
-        const segmentCustomers =
-          assertStringArray(segment.fields.customers) || [];
-        const segmentDevices = assertStringArray(segment.fields.devices) || [];
-        const segmentRegions = assertStringArray(segment.fields.regions) || [];
-
-        // Merge segment dimensions with existing dimensions
-        if (availabilityCustomers.length === 0) {
-          availabilityCustomers = [...segmentCustomers];
-        }
-        if (availabilityDevices.length === 0) {
-          availabilityDevices = [...segmentDevices];
-        }
-        if (availabilityRegions.length === 0) {
-          availabilityRegions = [...segmentRegions];
-        }
-      }
-    });
-  }
-
-  // Helper function to check if any requested dimension matches any available dimension (case insensitive)
-  const hasMatchingDimension = (
-    requested: string[],
-    available: string[],
-    dimensionType: string,
-  ) => {
-    if (requested.length === 0) return true; // No dimensions requested means no restriction
-    if (available.length === 0) return false; // No dimensions available means no access
-
-    // Get dimension names from IDs for comparison
-    const getNameFromId = (id: string, type: string) => {
-      // Type-safe access to airtable data arrays
-      let dataArray: AirtableRecord<FieldSet>[] | undefined;
-      if (type === "customerTypes")
-        dataArray = airtableData.dimensions?.customerTypes;
-      else if (type === "deviceTypes")
-        dataArray = airtableData.dimensions?.deviceTypes;
-      else if (type === "regions") dataArray = airtableData.dimensions?.regions;
-      else dataArray = undefined;
-
-      if (!dataArray) return id;
-
-      const record = dataArray.find((item) => item.id === id);
-      if (!record) return id;
-
-      // Try different possible name fields - prioritize slug for machine-readable identifiers
-      const name =
-        record.fields.slug || record.fields.name || record.fields.title || id;
-      return typeof name === "string" ? name.toLowerCase() : id;
-    };
-
-    const availableNames = available.map((id) =>
-      getNameFromId(id, dimensionType),
-    );
-    return requested.some((reqDim) =>
-      availableNames.some(
-        (availName) => availName.includes(reqDim) || reqDim.includes(availName),
-      ),
-    );
-  };
-
-  // Check all dimensions - all must match for strict availability
-  const customerMatch = hasMatchingDimension(
-    requestedDimensions.customerTypes,
-    availabilityCustomers,
-    "customerTypes",
-  );
-
-  const deviceMatch = hasMatchingDimension(
-    requestedDimensions.deviceTypes,
-    availabilityDevices,
-    "deviceTypes",
-  );
-
-  const regionMatch = hasMatchingDimension(
-    requestedDimensions.regions,
-    availabilityRegions,
-    "regions",
-  );
-
-  return customerMatch && deviceMatch && regionMatch;
-};
-
-export const filterContentByAvailability = (
-  contentAvailabilityIds: string[],
-  requestedDimensions: AvailabilityDimensions,
-  airtableData: Airtables,
-): boolean => {
-  let availabilityIdsToCheck = contentAvailabilityIds;
-
-  // If no availability records, use the default availability record
-  if (!contentAvailabilityIds || contentAvailabilityIds.length === 0) {
-    const defaultAvailability = airtableData.availability?.find(
-      (avail) => avail.fields.default === true,
-    );
-
-    if (defaultAvailability) {
-      availabilityIdsToCheck = [defaultAvailability.id];
-    } else {
-      return false; // No availability and no default means no access
-    }
-  }
-
-  // Content is available if ANY of its availability records match the requested dimensions
-  const hasMatch = availabilityIdsToCheck.some((availabilityId) => {
-    const availabilityRecord = airtableData.availability?.find(
-      (avail) => avail.id === availabilityId,
-    );
-
-    if (!availabilityRecord) return false;
-
-    return checkAvailabilityMatch(
-      availabilityRecord,
-      requestedDimensions,
-      airtableData,
-    );
-  });
-
-  return hasMatch;
 };
 
 // Generic sorting function that handles different property types
